Move webapis examples into its own module
