+++ examples/java/webapis/build.gradle
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * License); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an AS IS BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins { id 'org.apache.beam.module' }
+applyJavaNature(
+        automaticModuleName: 'org.apache.beam.examples.webapis'
+)
+
+description = "Apache Beam :: Examples :: Java :: WebApis"
+ext.summary = "Examples how to read from and write to Web APIs"
+
+dependencies {
+    implementation project(path: ":sdks:java:core", configuration: "shadow")
+    implementation project(":sdks:java:io:rrio")
+
+    implementation platform(library.java.google_cloud_platform_libraries_bom)
+    implementation library.java.google_cloud_vertexai
+
+    implementation library.java.slf4j_api
+    implementation library.java.commons_math3
+    implementation 'org.knowm.xchart:xchart:3.8.7'
+
+    runtimeOnly project(path: ":runners:direct-java", configuration: "shadow")
+
+    testImplementation project(path: ":runners:direct-java", configuration: "shadow")
+}
