+++ examples/java/webapis/src/main/java/org/apache/beam/examples/webapis/GeminiAIClient.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.snippets.transforms.io.webapis;
+package org.apache.beam.examples.webapis;
 
 import static org.apache.beam.sdk.util.Preconditions.checkStateNotNull;
