+++ examples/java/webapis/src/main/java/org/apache/beam/examples/webapis/ImageRequest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.snippets.transforms.io.webapis;
+package org.apache.beam.examples.webapis;
 
 import com.google.auto.value.AutoValue;
 import java.io.Serializable;
