+++ examples/java/webapis/src/test/java/org/apache/beam/examples/webapis/GeminiAIExampleTest.java
@@ -15,30 +15,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.snippets.transforms.io.webapis;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.notNullValue;
+package org.apache.beam.examples.webapis;
 
 import com.google.common.collect.ImmutableList;
+
+import java.io.IOException;
 import java.util.List;
-import org.apache.beam.io.requestresponse.Result;
-import org.apache.beam.sdk.testing.PAssert;
-import org.apache.beam.sdk.testing.TestPipeline;
-import org.apache.beam.sdk.transforms.Count;
-import org.junit.Rule;
+import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Tests for {@link UsingHttpClientExample}. */
+/** Tests for {@link GeminiAIExample}. */
 @RunWith(JUnit4.class)
-public class UsingHttpClientExampleTest {
-  @Rule public final TestPipeline pipeline = TestPipeline.create();
+public class GeminiAIExampleTest {
+  private static final GeminiAIOptions OPTIONS =
+      PipelineOptionsFactory.create().as(GeminiAIOptions.class);
+
+  static {
+    OPTIONS.setLocation("us-central1");
+    OPTIONS.setProjectId("apache-beam-testing");
+  }
 
   @Test
-  public void testReadFromGetEndpointExample() {
+  public void testWhatIsThisImage() {
     List<String> urls =
         ImmutableList.of(
             "https://storage.googleapis.com/generativeai-downloads/images/cake.jpg",
@@ -47,20 +47,11 @@ public class UsingHttpClientExampleTest {
             "https://storage.googleapis.com/generativeai-downloads/images/dog_form.jpg",
             "https://storage.googleapis.com/generativeai-downloads/images/factory.png",
             "https://storage.googleapis.com/generativeai-downloads/images/scones.jpg");
-    Result<ImageResponse> result =
-        UsingHttpClientExample.readFromGetEndpointExample(urls, pipeline);
-    PAssert.that(result.getFailures()).empty();
-    PAssert.thatSingleton(result.getResponses().apply(Count.globally())).notEqualTo(0L);
-    PAssert.that(result.getResponses())
-        .satisfies(
-            itr -> {
-              for (ImageResponse response : itr) {
-                assertThat(response, notNullValue());
-                assertThat(response.getData().isEmpty(), is(false));
-              }
-              return null;
-            });
+    GeminiAIExample.whatIsThisImage(urls, OPTIONS);
+  }
 
-    pipeline.run();
+  @Test
+  public void testIdentifyStats() throws IOException {
+    GeminiAIExample.identifyStats(OPTIONS);
   }
 }
